+++ internal/test-utils-core/src/commonMain/kotlin/com/github/panpf/sketch/test/utils/UriInvalidCondition.kt
@@ -1,0 +1,11 @@
+package com.github.panpf.sketch.test.utils
+
+import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.request.UriInvalidException
+import com.github.panpf.sketch.state.ErrorStateImage
+
+data object UriInvalidCondition : ErrorStateImage.Condition {
+
+    override fun accept(request: ImageRequest, throwable: Throwable?): Boolean =
+        throwable is UriInvalidException && (request.uri.isEmpty() || request.uri.isBlank())
+}