+++ sketch-animated/src/androidInstrumentedTest/kotlin/com/github/panpf/sketch/animated/android/test/request/internal/ImageRequestExecuteAnimatedTest.kt
@@ -35,7 +35,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class ImageRequestExecuteTest {
+class ImageRequestExecuteAnimatedTest {
 
     @Test
     fun testDisallowAnimatedImage() {