+++ sketch-core/src/androidMain/kotlin/com/github/panpf/sketch/state/ErrorStateImage.android.kt
@@ -1,21 +1,36 @@
 package com.github.panpf.sketch.state
 
 import com.github.panpf.sketch.drawable.DrawableEqualizer
-import com.github.panpf.sketch.state.ErrorStateImage.UriEmptyCondition
+import com.github.panpf.sketch.state.ErrorStateImage.Condition
+import com.github.panpf.sketch.util.IntColor
 
 
 /**
  * Add a StateImage dedicated to the empty uri error
  */
-fun ErrorStateImage.Builder.uriEmptyError(emptyDrawable: DrawableEqualizer): ErrorStateImage.Builder =
-    apply {
-        addState(UriEmptyCondition to DrawableStateImage(emptyDrawable))
-    }
+fun ErrorStateImage.Builder.addState(
+    condition: Condition,
+    drawable: DrawableEqualizer
+): ErrorStateImage.Builder = apply {
+    addState(condition, DrawableStateImage(drawable))
+}
 
 /**
  * Add a StateImage dedicated to the empty uri error
  */
-fun ErrorStateImage.Builder.uriEmptyError(emptyResId: Int): ErrorStateImage.Builder =
-    apply {
-        addState(UriEmptyCondition to DrawableStateImage(emptyResId))
-    }
+fun ErrorStateImage.Builder.addState(
+    condition: Condition,
+    resId: Int
+): ErrorStateImage.Builder = apply {
+    addState(condition, DrawableStateImage(resId))
+}
+
+/**
+ * Add a StateImage dedicated to the empty uri error
+ */
+fun ErrorStateImage.Builder.addState(
+    condition: Condition,
+    intColor: IntColor
+): ErrorStateImage.Builder = apply {
+    addState(condition, ColorDrawableStateImage(intColor))
+}